feat: adds derived resources and moves highwater from postprocessing
